fix(v2): isolates feature and credits cra
